requirements.txt
torch==2.0.0
torchvision==0.15.1
transformers==4.29.0
numpy==1.23.0
Pillow==9.5.0
scipy==1.10.0
